+++ animator/src/cs3500/animator/model/IAnimatorModel.java
@@ -50,7 +50,7 @@ public interface IAnimatorModel extends IModelView{
    * @param time the time we want the state at
    * @return the text representation of the animator
    */
-  String toString(float time);
+  String toString(int time);
 
   /**
    * Returns the state of the animator as a collection of states.
@@ -58,7 +58,7 @@ public interface IAnimatorModel extends IModelView{
    * @param time the time we want the state at
    * @return a list of states representing the shapes at the given time
    */
-  List<IShape> getState(float time);
+  List<IShape> getState(int time);
 
   /**
    * Returns a list of the shapes by name in the animator.model.
